+++ frontend/src/utils/consts.ts
@@ -147,6 +147,7 @@ export const TEXTS_HEBREW = {
     LANDED_FLIGHTS: 'נחתו',
     NO_FLIGHTS_IN_THIS_STATE: 'אין טיסות במצב זה',
     RESEND_NOTIFICATION: 'שלח מחדש',
+    RESEND_NOTIFICATION_CONFIRMATION: 'האם אתה בטוח שברצונך לשלוח מחדש את ההודעה?',
 }