+++ RELEASE-NOTES.md
@@ -6,6 +6,10 @@
 
 - Adds a button in the notifications page in the UI to resend the notification in case it failed to send.
 
+### Enhancements
+
+- Adds the properties to the `NotificationUpdateSchema` in the backend in order to allow updating the state of the notification.
+
 ## v3.0.2
 
 ### Enhancements